+++ lib/grape_oas/api_model_builder.rb
@@ -20,6 +20,7 @@ module GrapeOAS
       @api.registered_schemas = build_registered_schemas(options[:models])
       @api.suppress_default_error_response = options[:suppress_default_error_response] || false
       @api.nullable_strategy = options[:nullable_strategy]
+      @api.array_use_braces = options[:array_use_braces] || false
 
       @namespace_filter = options[:namespace]
       @apis = []