+++ lib/grape_oas/exporter/oas3/operation.rb
@@ -11,10 +11,13 @@ module GrapeOAS
         # OAS3-specific fields: parameters (no body), requestBody, responses
         def build_version_specific_fields
           strategy = @options[:nullable_strategy] || Constants::NullableStrategy::KEYWORD
+          braces = @options[:array_use_braces] || false
 
           {
-            "parameters" => Parameter.new(@op, @ref_tracker, nullable_strategy: strategy).build,
-            "requestBody" => RequestBody.new(@op.request_body, @ref_tracker, nullable_strategy: strategy).build,
+            "parameters" => Parameter.new(@op, @ref_tracker, nullable_strategy: strategy,
+                                                             array_use_braces: braces,).build,
+            "requestBody" => RequestBody.new(@op.request_body, @ref_tracker, nullable_strategy: strategy,
+                                                                             array_use_braces: braces,).build,
             "responses" => Response.new(@op.responses, @ref_tracker, nullable_strategy: strategy).build
           }
         end