+++ lib/grape_oas/exporter/oas3/paths.rb
@@ -12,6 +12,7 @@ module GrapeOAS
         def build_operation(operation)
           Operation.new(operation, @ref_tracker,
                         nullable_strategy: @options[:nullable_strategy] || Constants::NullableStrategy::KEYWORD,
+                        array_use_braces: @options[:array_use_braces] || false,
                         suppress_default_error_response: @options[:suppress_default_error_response],).build
         end
       end